requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
numpy==2.2.6
faiss-cpu==1.12.0
sentence-transformers==2.6.1
transformers==4.41.2
huggingface-hub==0.23.5
rank-bm25==0.2.2
pypdf==5.9.0
tqdm==4.67.1
openai==1.102.0
